==> pandas/1-series.ipynb <==
import pandas as pd

dados = [10,20,30,40,50]
serie = pd.Series(dados)
serie

type(serie)

# Criando uma Series com dados de vendas mensais
vendas_mensais = pd.Series(
    [2005,1745,2500,3000,2800,3500,4000,5500,4897, 9900,5666,7854],
    index=['Jan','Fev','Mar','Abr','Mai','Jun','Jul','Ago','Set','Out','Nov','Dez'] 
)

vendas_mensais

# Calculando e exibir a média de vendas
media_vendas = vendas_mensais.mean()
media_vendas

# Identificando o mes com maior venda
mes_maior_vendas = vendas_mensais.idxmax()
mes_maior_vendas

# Exibir as vendas acima da média
vendas_acima_media = vendas_mensais[vendas_mensais > media_vendas]
vendas_acima_media
