==> pandas/2-dataframe.ipynb <==
import pandas as pd

dados = {
    'Nome': ["Alice", "Maquerle", "Marcelo"],
    'Idade': [5,32,34],
    'Cidade': ["Vitória", "Cachoeiro", "Serra"]
}

dados

df = pd.DataFrame(dados)
df

df.head()

dados_notas = {
    "Nome": ["Alice", "Bob", "Charlie", "David", "Eva"],
    "Matematica": [85,86,99,60,57],
    "Fisica": [78, 66, 95, 59, 67],
    "Quimica": [87, 76, 65, 70, 87]
}

df_notas = pd.DataFrame(dados_notas)


df_notas

# Calculando a média das notas por aluno
df_notas['Média'] = df_notas[['Matematica', 'Fisica', 'Quimica']].mean(axis=1)

df_notas

#Identificar alunos aprovados com nota boa (média >= 82.5)
alunos_aprovados = df_notas[df_notas['Média'] >= 82.5]
alunos_aprovados

# Exibir a média e os alunos aprovados
df_notas[['Nome', 'Média']]

alunos_aprovados[['Nome', 'Média']]
