==> pandas/5-importar_csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv('../dados/produtos.csv')
df_produtos

df_produtos.head()

df_produtos.tail()

df_produtos.describe()

df_produtos.info()

df_produtos.shape

df_produtos.dtypes

df_produtos.sort_values(by='Quantidade', ascending=False)

df_produtos['Categoria'].value_counts()
