==> pandas/7-criar_planilha.ipynb <==
import pandas as pd

dados = {
    'Times': ['Flamengo', 'Palmeiras', 'Vasco', 'Fluminense', 'Santos'],
    'Partidas_Jogadas': [10,10,10,10,10],
    'Vitórias': [7,5,6,4,3],
    'Empates': [5,2,3,4,5],
    'Derrotas': [1,2,3,3,4],
    'Gols_Pro': [22,18,16,15,13],
    'Gols_Contra': [10,12,15,16,20]
}

df = pd.DataFrame(dados)
df

df['Saldo_Gols'] = df['Gols_Pro'] - df['Gols_Contra']
df

df['Pontos'] = df['Vitórias'] * 3 + df['Empates'] + 1
df

df = df.sort_values(by='Pontos', ascending=False)
df

df.to_excel('../dados/campeonato_futebol2.xlsx', index=False)
