==> pandas/8-analise_planilha.ipynb <==
import pandas as pd

df = pd.read_excel("../dados/campeonato_futebol.xlsx")
df

df.dtypes

df.describe()

df.head(4)

df.tail(4)

# 1 - análise de desempenho por tim
desempenho_info = df[['Times', 'Pontos', 'Vitorias', 'Saldo_Gols']].set_index('Times')
desempenho_info

# 2 - Time com mais vitória e menos vitoria
time_mais_vitoria = df[df['Vitorias'] == df['Vitorias'].max()]
time_mais_vitoria

time_menos_vitoria = df[df['Vitorias'] == df['Vitorias'].min()]
time_menos_vitoria

# 3 - Distribuição de Gols Pró
df['Gols_Pro'].value_counts()

# 4 - Relçao entreo gols pró e contas
correlacao = df['Gols_Pro'].corr(df['Pontos'])
correlacao
