==> pandas/12-db.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

engine = create_engine("sqlite:///ecommerce.db")

df = pd.read_csv("../dados/produtos.csv")
df.head()

try:
    df.to_sql('produtos', con=engine, if_exists='replace', index=False)
    print("Dados exportados com sucesso!")
except Exception as e:
    print(f"Ocorreu um erro ao exportar os dados {e}")   

# 1 - Total de vendas por categoria
consulta_1 = "SELECT Categoria, SUM(Quantidade) AS Total_Qtd_Vendida FROM produtos GROUP BY Categoria"
resultado_1 = pd.read_sql_query(consulta_1, engine)
resultado_1

#2 - Produto com mais vendas
consulta_2 = "SELECT Produto, SUM(Quantidade) AS Total_Qtd_Vendida FROM produtos GROUP BY Produto ORDER BY Total_Qtd_Vendida DESC LIMIT 1"
resultado_2 = pd.read_sql_query(consulta_2, engine)
resultado_2

# 3 - Total de Vendas por Produto
consulta_3 = "SELECT Produto, SUM(Quantidade) AS Total_Qtd_Vendida FROM produtos GROUP BY Produto ORDER BY Total_Qtd_Vendida DESC"
resultado_3 = pd.read_sql_query(consulta_3, engine)
resultado_3

# 4 - Média de preço por Produto
consulta_4 = "SELECT Produto, AVG(Preço) AS Media_Preco FROM produtos GROUP BY Produto"
resultado_4 = pd.read_sql_query(consulta_4, engine)
resultado_4
